backend/requirements.txt
fastapi==0.95.1
uvicorn==0.23.1
sqlalchemy==2.0.0
pymysql==1.1.1
python-dotenv==0.21.1
pydantic==1.10.7
cryptography==44.0.2
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.4.0
email-validator==2.2.0